# Conflicts:
#	src/main/java/com/codeup/phaserun/controllers/ProfileController.java
#	src/main/resources/static/css/profile.css
#	src/main/resources/templates/partials/navbar.html
#	src/main/resources/templates/users/profile.html
#	src/main/resources/templates/visitors/home.html

+++ src/main/java/com/codeup/phaserun/repositories/RaceRepository.java
@@ -1,15 +1,19 @@
 package com.codeup.phaserun.repositories;
 
 import com.codeup.phaserun.models.Race;
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
 public interface RaceRepository extends JpaRepository<Race, Long> {
-    Race findById(int id);
+    Race findById(long id);
     boolean existsByRaceId(String raceId);
+
     Race findByRaceId(String raceId);
 }
